Changes to the code
Made changes to the code. Started off by deleting place holder classes and adding in actual classes that should be needed, might need to add more in as we continue to work. Started Implementation on the Client side for the checkers game. Still need to go back in and add in certain parts to the CheckersGUI and CheckersClient and possibly more classes, but this is a start